% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qa_sites_report.R
\name{generate_qa_sites_report}
\alias{generate_qa_sites_report}
\title{Generate QA report for surveys and sites}
\usage{
generate_qa_sites_report(odkc_ex, odkc_tf, year)
}
\description{
Generate QA report for surveys and sites
}
\examples{
\dontrun{
generate_qa_sites_report(odkc_ex, odkc_tf, wastd_data_yr)
}
}
