% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qa_users_report.R
\name{generate_qa_users_report}
\alias{generate_qa_users_report}
\title{Generate QA report for user mapping}
\usage{
generate_qa_users_report(odkc_ex, user_mapping, year)
}
\description{
Generate QA report for user mapping
}
\examples{
\dontrun{
generate_qa_users_report(odkc_ex, user_mapping, wastd_data_yr)
}
}
