% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qa_users_report_2019.R
\name{generate_qa_users_report_2019}
\alias{generate_qa_users_report_2019}
\title{Generate QA report for user mapping}
\usage{
generate_qa_users_report_2019(odkc_ex, user_mapping, year)
}
\description{
Generate QA report for user mapping
}
\examples{
\dontrun{
generate_qa_users_report_2019(odkc_ex, user_mapping, wastd_data_yr)
}
}
