% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qa_users_report.R
\name{generate_qa_users_report_w2}
\alias{generate_qa_users_report_w2}
\title{Generate QA report for W2 user mapping}
\usage{
generate_qa_users_report_w2(
  user_mapping,
  w2_yr,
  w2_data,
  verbose = wastdr::get_wastdr_verbose()
)
}
\description{
Generate QA report for W2 user mapping
}
\examples{
\dontrun{
generate_qa_users_report(odkc_ex, user_mapping, wastd_data_yr)
}
}
