% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_wastd_reports.R
\name{generate_wastd_reports}
\alias{generate_wastd_reports}
\title{Generate reports from WAStD turtle data}
\usage{
generate_wastd_reports(wastd_data)
}
\description{
Generate reports from WAStD turtle data
}
\examples{
\dontrun{
drake::loadd("wastd_data")
rmarkdown::render(
here::here("vignettes/wastd.Rmd"),
params = list(area_name = "Delambre Island"),
output_file = here::here("vignettes/wastd_del.html"))
}
}
