% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_svs_sve_as_wastd_surveys.R
\name{get_media_file}
\alias{get_media_file}
\title{Turn a filename into a file object if exists else return NULL.}
\usage{
get_media_file(value)
}
\description{
Turn a filename into a file object if exists else return NULL.
}
