% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_todays_folder_sharepoint.R
\name{get_todays_folder_sharepoint}
\alias{get_todays_folder_sharepoint}
\title{Get a folder with the current date from a SharePoint site's "General" drive}
\usage{
get_todays_folder_sharepoint(team_name)
}
\arguments{
\item{team_name}{An existing Team name, e.g. "TurtleData", "TurtlesRioTinto",
"TurtlesC4H", "TurtlesPendoley".}
}
\value{
The Microsoft365R R6 object for a shared folder named
\code{General/YYYY-MM-DD} with today's date.
}
\description{
Get a folder with the current date from a SharePoint site's "General" drive
}
