% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_unknown_user.R
\name{get_user_area}
\alias{get_user_area}
\title{Return a comma-separated list of area_names for a given username from odkc_data}
\usage{
get_user_area(odkc_data, username)
}
\description{
Return a comma-separated list of area_names for a given username from odkc_data
}
\examples{
\dontrun{
drake::loadd("odkc_ex")
get_user_area(odkc_ex, "Janani Nakulan")
get_user_area(odkc_ex, "Kloe Ams")
get_user_area(odkc_ex, "jessica mcglashan, ashley bachert")
get_user_area(odkc_ex, "Melanie Lambert, Elena Miller")
}
}
