% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_user_mapping.R
\name{make_user_mapping}
\alias{make_user_mapping}
\title{Generate a WAStD user mapping from a given list of ODKC data.}
\usage{
make_user_mapping(odkc_data, wastd_users)
}
\arguments{
\item{odkc_data}{The output of \code{wastdr::download_all_odkc_turtledata_2019}.}

\item{wastd_users}{A tibble of WAStD users}
}
\description{
The matching is done by
\href{http://varianceexplained.org/fuzzyjoin/reference/stringdist_join.html}{\code{fuzzyjoin::stringdist_left_join}}
using the Jaro-Winker distance between the ODKC username and the individual
WAStD \code{name}, \code{username} and \code{aliases}.
}
\details{
The field \code{aliases} is where the magic happens. \code{make_user_mapping} matches
against each alias after separating the aliases by comma.
This way, we can add the exact misspelling of an ODK Collect username as
a new alias, and get a 100\% match for it.

Extract all unique reporter names from odkc data.
Extract relevant user names from WAStD users.
Map most likely match and export to CSV.
External QA: review mapping, update WAStD user aliases to improve the
user matching process. Re-run until optimized, edit CSV to improve match.
Return a named list containing the mapping of odkc_reporter and wastd_user_id.
}
