% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_user_mapping_w2.R
\name{make_user_mapping_w2}
\alias{make_user_mapping_w2}
\title{Generate a WAStD user mapping from a given turtle tagging DB export}
\usage{
make_user_mapping_w2(w2_data, wastd_users)
}
\arguments{
\item{w2_data}{The output of \code{wastdr::download_w2_data}.}

\item{wastd_users}{A tibble of WAStD users.}
}
\value{
A tibble of legacy username and user ID with the respective best
match of a WAStD user.
}
\description{
The matching is done by
\href{http://varianceexplained.org/fuzzyjoin/reference/stringdist_join.html}{\code{fuzzyjoin::stringdist_left_join}}
using the Jaro-Winker distance between the ODKC username and the individual
WAStD \code{name}, \code{username} and \code{aliases} of current active WAStD Users.
}
\details{
The field \code{aliases} is where the magic happens.
\code{make_user_mapping_w2} matches against each alias after separating the
aliases by comma.
This way, we can add the exact misspelling of an ODK Collect username as
a new alias, and get a 100\% match for it.
}
