% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_unknown_user.R
\name{map_unknown_user}
\alias{map_unknown_user}
\title{Generate a map of odkc_ex$tracks for a given username}
\usage{
map_unknown_user(odkc_data, odkc_username)
}
\description{
Generate a map of odkc_ex$tracks for a given username
}
\examples{
\dontrun{
drake::loadd("odkc_ex")
map_unknown_user(odkc_ex, "Janani Nakulan")
}
}
