% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_wastd_wamtram_sites.R
\name{map_wastd_wamtram_sites}
\alias{map_wastd_wamtram_sites}
\title{Map WAStD and WAMTRAM sites}
\usage{
map_wastd_wamtram_sites(data_wastd, data_wamtram)
}
\arguments{
\item{data_wastd}{An object of class \code{wastd_data}.}

\item{data_wamtram}{An object of class \code{wamtram_data}}
}
\value{
A leaflet map
}
\description{
This map is useful to compare and update WAStD with and from WAMTRAM sites.
}
\examples{
\dontrun{
data("wastd_data")
data("wamtram_data")
map_wastd_wamtram_sites(wastd_data, wamtram_data)
}
}
