% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_as_survey_media.R
\name{odkc_as_survey_media}
\alias{odkc_as_survey_media}
\title{ODKC SiteVisitStart/End as WAStD SurveyMediaObservations}
\usage{
odkc_as_survey_media(data)
}
\arguments{
\item{data}{ODKC Data, e.g. \code{odkc_data}.}
}
\value{
A tibble of records which can be uploaded ONE BY ONE to WAStD's
SurveyMediaAttachment API as per function examples.
The field "attachment" has to be turned into an \code{httr::upload_file} before
upload.
}
\description{
ODKC SiteVisitStart/End as WAStD SurveyMediaObservations
}
\examples{
\dontrun{
wastdr::wastdr_setup(api_url = Sys.getenv("WASTDR_API_DEV_URL"),
                     api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"))
drake::loadd(odkc_ex)
x <- odkc_as_survey_media(odkc_ex)
upload_survey_media(x)


# Outside R, using curl (with correct auth token)
curl -i -X POST
  -H 'Authorization: Token c5575c09f6a0171668f31d5dd5013f02658668bd'
  -F 'attachment=@my_photo.jpg'
  -F survey_source=odk
  -F survey_source_id=5f027f76-7276-11eb-8ce8-e9c53fd9c3f2
  -F source=2
  -F source_id=5f027f76-7276-11eb-8ce8-e9c53fd9c3f2-photo-start
  -F title='Photo site conditions at start of survey'
  http://localhost:8220/api/1/survey-media-attachments/
}
}
