% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_svs_sve_as_wastd_surveys.R
\name{odkc_svs_sve_as_wastd_surveys}
\alias{odkc_svs_sve_as_wastd_surveys}
\title{Reconstruct WAStD Surveys from ODKC Site Visit Start/End}
\usage{
odkc_svs_sve_as_wastd_surveys(svs, sve, user_mapping)
}
\arguments{
\item{svs}{ODKC Site Visit Start, e.g. \verb{data("odkc_data"); odkc_data$svs}}

\item{sve}{ODKC Site Visit End, e.g. \verb{data("odkc_data"); odkc_data$sve}}
}
\value{

}
\description{
Reconstruct WAStD Surveys from ODKC Site Visit Start/End
}
\details{
Surveys are reconstructed from Site Visit Start records.
Survey end details are populated from Site Visit End records, where matching
by \code{device_id}, \code{calendar_date_awst}, and \code{site_id}.

Sources of error:
\itemize{
\item SVS forgotten to record
\item SVE forgotten to record
\item SVS recorded outside known site boundaries (WAStD Area),
\code{site_id} mismatch, svs$site_id empty
\item SVE recorded outside known site boundaries (WAStD Area),
\code{site_id} mismatch, sve$site_id empty
\item Changed devices mid survey (\code{device_id} mismatch, likely reporter mismatch,
see end_comments)
}

Things to provide as QA reports outside this function:
\itemize{
\item Stray SVE without matching SVS
\item SVS joined to SVE only by device_id and calendar_date_awst minus "good"
surveys (location out of bounds)
\item SVS joined to SVE only by calendar_date_awst and site_id minus "good"
surveys (device swapped?)
}
}
\examples{
wastdr_msg_info("examples go here")
}
