% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tracks_as_media.R
\name{odkc_tracks_as_media}
\alias{odkc_tracks_as_media}
\title{ODKC Tracks as WAStD MediaObservations, WIP.}
\usage{
odkc_tracks_as_media(data)
}
\description{
ODKC Tracks as WAStD MediaObservations, WIP.
}
\examples{
\dontrun{
wastdr::wastdr_setup(api_url = Sys.getenv("WASTDR_API_DEV_URL"),
                     api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"))
drake::loadd(odkc_ex)
x <- odkc_tracks_as_media(odkc_ex$tracks)
# Doesn't work yet:
wastdr::wastd_POST(as.list(x[1,]), serializer="media-attachments", encode="multipart")
}
}
