% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_as_wastd_nesttagobs.R
\name{odkc_tt_as_wastd_nesttagobs}
\alias{odkc_tt_as_wastd_nesttagobs}
\title{Transform odkc_data$tt into WAStD NestTagObservations.}
\usage{
odkc_tt_as_wastd_nesttagobs(data)
}
\arguments{
\item{data}{A tibble of turtle tagging data, e.g. \code{odkc_data$tt}.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("nest-tag-observations")}
}
\description{
Transform odkc_data$tt into WAStD NestTagObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tt_as_wastd_nesttagobs(odkc_data$tt)
x \%>\% wastdr::wastd_POST("nest-tag-observations", api_url = au, api_token = at)
}
}
