% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_as_wastd_turtledmg.R
\name{odkc_tt_as_wastd_turtledmg}
\alias{odkc_tt_as_wastd_turtledmg}
\title{Transform odkc_data$tt into WAStD TurtleDamageObservations (scars, missed
tags)}
\usage{
odkc_tt_as_wastd_turtledmg(data)
}
\arguments{
\item{data}{A tibble of turtle tagging data, e.g. \code{odkc_data$tt}.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("turtle-damage-observations")}
}
\description{
Transform odkc_data$tt into WAStD TurtleDamageObservations (scars, missed
tags)
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tt_as_wastd_turtledmg(odkc_data$tt)
x \%>\% wastdr::wastd_POST("turtle-damage-observations",
  api_url = au, api_token = at)
}
}
