% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_as_wastd_turtlemorph.R
\name{odkc_tt_as_wastd_turtlemorph}
\alias{odkc_tt_as_wastd_turtlemorph}
\title{Transform odkc_data$tt into WAStD TurtleMorphometricObservations.}
\usage{
odkc_tt_as_wastd_turtlemorph(data, user_mapping)
}
\arguments{
\item{data}{A tibble of Turtle Tagging records, e.g. \code{odkc_data$tt}.}

\item{user_mapping}{.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("turtle-morphometrics")}
}
\description{
Transform odkc_data$tt into WAStD TurtleMorphometricObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
data("wastd_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
drake::loadd("user_mapping")
tmorph <- odkc_tt_as_wastd_turtlemorph(odkc_data$tt, user_mapping)

res <- tmorph \%>\%
  wastdr::wastd_POST("turtle-morphometrics", api_url = au, api_token = at)
}
}
