% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_as_wastd_turtlenestobs.R
\name{odkc_tt_as_wastd_turtlenestobs}
\alias{odkc_tt_as_wastd_turtlenestobs}
\title{Transform odkc_data$tt into WAStD TurtleNestObservations (egg counts).}
\usage{
odkc_tt_as_wastd_turtlenestobs(data)
}
\arguments{
\item{data}{A tibble of turtle tagging data, e.g. \code{odkc_data$tt}.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("turtle-nest-excavations")}
}
\description{
Transform odkc_data$tt into WAStD TurtleNestObservations (egg counts).
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tt_as_wastd_turtlenestobs(odkc_data$tt)
x \%>\% wastd_POST("turtle-nest-excavations", api_url = au, api_token = at)
}
}
