% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_log_as_wastd_loggerobs.R
\name{odkc_tt_log_as_wastd_loggerobs}
\alias{odkc_tt_log_as_wastd_loggerobs}
\title{Transform odkc_data$tt_log into WAStD LoggerObservations.}
\usage{
odkc_tt_log_as_wastd_loggerobs(data)
}
\arguments{
\item{data}{A tibble of logger observations from turtle tagging,
e.g. \code{odkc_data$tt_log}.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("logger-observations")}
}
\description{
Transform odkc_data$tt_log into WAStD LoggerObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
data("wastd_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
x <- odkc_tt_as_wastd_loggerobs(odkc_data$tt_log)
x \%>\% wastd_POST("logger-observations", api_url = au, api_token = at)
}
}
