% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odkc_tt_tag_as_wastd_tagobs.R
\name{odkc_tt_tag_as_wastd_tagobs}
\alias{odkc_tt_tag_as_wastd_tagobs}
\title{Transform odkc_data into WAStD TagObservations.}
\usage{
odkc_tt_tag_as_wastd_tagobs(data, user_mapping)
}
\arguments{
\item{data}{An object of class odkc_data, e.g. \code{odkc_data}.}

\item{user_mapping}{The output of \code{\link{make_user_mapping}()},
containing at least columns \code{pk} (WAStD User PK) and
\code{odkc_username} (as written in odkc data field \code{reporter}).}
}
\value{
A tibble suitable to \code{\link{wastd_POST}("tag-observations")}.
}
\description{
Transform odkc_data into WAStD TagObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
data("wastd_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
drake::loadd("user_mapping")
x <- odkc_tt_tag_as_wastd_tagobs(odkc_data, user_mapping)
x \%>\% wastdr::wastd_POST("tag-observations", api_url = au, api_token = at)
}
}
