% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_time.R
\name{repair_time}
\alias{repair_time}
\title{Repair a time offset}
\usage{
repair_time(
  dt,
  ref_was = "2010-03-22 20:08:00 +08",
  ref_should = "2020-12-28 07:08:00 +08",
  tz = ruODK::get_default_tz()
)
}
\arguments{
\item{dt}{A datetime to be corrected by a time offset in format readable by
\code{lubridate::ymd_hms()}, containing the timezone.
E.g. "2010-03-22 20:23:47 +08".}

\item{ref_was}{A reference datetime for an uncorrected date in a format
readable by \code{lubridate::ymd_hms()}, containing the timezone.
Default: "2010-03-22 20:08:00 +08".}

\item{ref_should}{A reference datetime for the corrected \code{ref_was} in a format
readable by \code{lubridate::ymd_hms()}, containing the timezone.
Default: "2020-12-28 07:08:00 +08".}

\item{tz}{The timezone for the corrected datetime,
default: \code{ruODK::get_default_tz()}.}
}
\value{
The given datetime corrected by the timedelta between \code{ref_was} and
\code{ref_should}.
}
\description{
Repair a time offset
}
\examples{
repair_time("2010-03-22 20:23:47 +08")
repair_time("2010-03-03 18:13:01 +08")
repair_time("2010-03-03 18:30:53 +08")
repair_time("2010-03-03 18:49:25 +08")
repair_time("2010-03-23 01:49:49 +00")

}
