% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_create_update_skip.R
\name{split_create_update_skip}
\alias{split_create_update_skip}
\title{Top level helper to split all Turtle Nesting Census data per skip logic}
\usage{
split_create_update_skip(odkc_prep, wastd_data)
}
\arguments{
\item{odkc_prep}{ODKC data transformed into WAStD format.}

\item{wastd_data}{Minimal WAStD data to inform skip logic.}
}
\value{
A list of results from the various uploads.
Each result is a \code{wastd_api_response} or a tibble of data (\verb{_skip}).
}
\description{
Encounters and their related observations are uploaded to WAStD:
}
\details{
\itemize{
\item New encounters will be created
\item Existing but unchanged (status=new) encounters will be updated
if \code{update_existing=TRUE}, else skipped.
\item Existing and value-added encounters (status > new) will be skipped.s
}
}
