% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_item_to_sharepoint.R
\name{upload_item_to_sharepoint}
\alias{upload_item_to_sharepoint}
\title{Upload one file or folder to a SharePoint drive}
\usage{
upload_item_to_sharepoint(
  local_file_path,
  dst,
  local_root_rel = "inst/reports"
)
}
\arguments{
\item{local_file_path}{The relative or absolute path to an existing local
file to upload.}

\item{drive_path}{A SharePoint folder, as returned by \code{get_item} or
\code{create_folder} run on a SharePoint Drive.}
}
\description{
Given a local file path, the file is uploaded to
}
