% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_survey_media.R
\name{upload_survey_media}
\alias{upload_survey_media}
\title{Upload a tibble of SurveyMediaAttachments to WAStD}
\usage{
upload_survey_media(
  x,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose(),
  upload = Sys.getenv("ODKC_IMPORT_UPLOAD_MEDIA", unset = TRUE)
)
}
\arguments{
\item{x}{A tibble of SurveyMediaAttachments, e.g. the output of
\code{odkc_as_survey_media()}.
Note that media files should be relative or absolute file paths, not
\code{httr::upload_file()} objects.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}

\item{upload}{Whether to upload media files.
Default: \code{Sys.getenv("ODKC_IMPORT_UPLOAD_MEDIA", unset=TRUE)}}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}
}
\description{
Upload a tibble of SurveyMediaAttachments to WAStD
}
\details{
WAStD's SurveyMediaAttachment endpoint uses a plain ModelViewSet
and cannot handle multiple uploads.
This function therefore uploads records individually.
If a file does not exist locally, the record is skipped and a warning is
emitted.
}
