% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_surveys.R
\name{upload_surveys}
\alias{upload_surveys}
\title{Upload a tibble of Surveys to WAStD}
\usage{
upload_surveys(
  x,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{x}{A tibble of Surveys, e.g. the output of
\code{odkc_svs_sve_as_wastd_surveys()}.
Note that attachments should be relative or absolute file paths, not
\code{httr::upload_file()} objects.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}
}
\description{
Upload a tibble of Surveys to WAStD
}
\details{
WAStD's MediaAttachment endpoint uses a plain ModelViewSet and cannot
handle multiple uploads. This function therefore uploads records individually.
If a file does not exist locally, the record is skipped and a warning is
emitted.
}
