% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_w2_to_wastd.R
\name{upload_w2_to_wastd}
\alias{upload_w2_to_wastd}
\title{Top level helper to upload all Turtle Tagging data to WAStD}
\usage{
upload_w2_to_wastd(
  data,
  update_existing = FALSE,
  api_url = wastdr::get_wastdr_api_url(),
  api_token = wastdr::get_wastdr_api_token(),
  verbose = wastdr::get_wastdr_verbose()
)
}
\arguments{
\item{data}{W2 data transformed into WAStD format and split into create,
update, skip.}

\item{update_existing}{Whether to update existing but unchanged records in
WAStD Default: FALSE.}

\item{api_url}{(character) The WAStD API URL,
default \code{\link{get_wastdr_api_url}}, see \code{\link{wastdr_setup}}}

\item{api_token}{(character) The WAStD API token,
default \code{\link{get_wastdr_api_token}}, see \code{\link{wastdr_setup}}}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}

\item{api_un}{(character) A WAStD API username,
default \code{\link{get_wastdr_api_un}}, see \code{\link{wastdr_setup}}}

\item{api_pw}{(character) A WAStD API password,
default \code{\link{get_wastdr_api_pw}}, see \code{\link{wastdr_setup}}}
}
\value{
A list of results from \code{wastdr::wastd_create_update_skip}.
}
\description{
Encounters and their related observations are uploaded to WAStD:
}
\details{
\itemize{
\item New encounters will be created
\item Existing but unchanged (status=new) encounters will be updated
if \code{update_existing=TRUE}, else skipped.
\item Existing and value-added encounters (status > new) will be skipped.s
}
}
\examples{
\dontrun{
wastdr::wastdr_setup(api_url = Sys.getenv("WASTDR_API_DEV_URL"),
                     api_token = Sys.getenv("WASTDR_API_DEV_TOKEN"))
wastdr::wastdr_setup(api_url = Sys.getenv("WASTDR_API_TEST_URL"),
                     api_token = Sys.getenv("WASTDR_API_TEST_TOKEN"))
wastdr::wastdr_setup(api_url = Sys.getenv("WASTDR_API_URL"),
                     api_token = Sys.getenv("WASTDR_API_TOKEN"))
drake::loadd("w2_up")
upload_odkc_to_wastd(w2_up, update_existing=TRUE, verbose=TRUE)
}
}
