% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_as_wastd.R
\name{w2_as_wastd}
\alias{w2_as_wastd}
\title{Transform W2 data into WAStD data}
\usage{
w2_as_wastd(data, user_mapping)
}
\arguments{
\item{data}{Data of class \code{wamtram_data}.}

\item{user_mapping}{The output of \code{make_user_mapping_w2()}.}
}
\value{
A tibble suitable for \code{split_create_update_skip()}.
}
\description{
This function calls one function to produce the data for each API endpoint.
}
\seealso{
Other wamtram: 
\code{\link{w2_as_wastd_ae}()}
}
\concept{wamtram}
