% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_as_wastd_ae.R
\name{w2_as_wastd_ae}
\alias{w2_as_wastd_ae}
\title{Transform W2 data into WAStD AnimalEncounters}
\usage{
w2_as_wastd_ae(data, user_mapping, verbose = wastdr::get_wastdr_verbose())
}
\arguments{
\item{data}{Data of class \code{wamtram_data}.}

\item{user_mapping}{The output of \code{make_user_mapping_w2()}.}

\item{verbose}{Whether to show debug messages.
Default: wastdr::get_wastdr_verbose()}
}
\value{
A tibble suitable to upload to the WAStD AnimalEncounters API endpoint.
}
\description{
Transform W2 data into WAStD AnimalEncounters
}
\seealso{
Other wamtram: 
\code{\link{w2_as_wastd}()}
}
\concept{wamtram}
