% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_dmg_as_wastd_turtledmg.R
\name{w2_dmg_as_wastd_turtledmg}
\alias{w2_dmg_as_wastd_turtledmg}
\title{Transform w2_data into WAStD TurtleDamageObservations.}
\usage{
w2_dmg_as_wastd_turtledmg(data)
}
\arguments{
\item{data}{Data of class \code{wamtram_data}.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("turtle-damage-observations")}
}
\description{
Transform w2_data into WAStD TurtleDamageObservations.
}
\examples{
\dontrun{
}
}
