% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_tag_as_wastd_tagobs.R
\name{w2_tag_as_wastd_tagobs}
\alias{w2_tag_as_wastd_tagobs}
\title{Transform odkc_data$mwi_tag into WAStD TagObservations.}
\usage{
w2_tag_as_wastd_tagobs(data, user_mapping)
}
\arguments{
\item{data}{Data of class \code{wamtram_data}.}

\item{user_mapping}{The output of \code{make_user_mapping_w2()}.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("tag-observations")}
}
\description{
Transform odkc_data$mwi_tag into WAStD TagObservations.
}
\examples{
\dontrun{
data("odkc_data", package = "wastdr")
au <- Sys.getenv("WASTDR_API_DEV_URL")
at <- Sys.getenv("WASTDR_API_DEV_TOKEN")
user_mapping <- NULL # see odkc_plan for actual user mapping
x <- odkc_mwi_tag_as_wastd_tagobs(odkc_ex$mwi_tag, user_mapping)
x \%>\% wastd_POST("tag-observations", api_url = au, api_token = at)
}
}
\seealso{
https://github.com/dbca-wa/wastd/blob/master/wastd/observations/utils.py#L1849
}
