% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/w2_tag_as_wastd_turtledmg.R
\name{w2_tag_as_wastd_turtledmg}
\alias{w2_tag_as_wastd_turtledmg}
\title{Transform w2_data into WAStD TurtleDamageObservations for tag scars.}
\usage{
w2_tag_as_wastd_turtledmg(data)
}
\arguments{
\item{data}{Data of class \code{wamtram_data}.}
}
\value{
A tibble suitable to
\code{\link{wastd_POST}("turtle-damage-observations")}
}
\description{
w2_data$obs_flipper_tags of tag_status M and M1 become
WAStD TurtleDamageObservations of damage_type tag-scar.
}
\examples{
\dontrun{
}
}
