% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_reports.R
\name{wastd_reports}
\alias{wastd_reports}
\title{Generate and disseminate WAStD reports}
\usage{
wastd_reports()
}
\description{
Writing to Google Sheets requires a Google Service Account Token to be
present at \verb{~/.config/gcloud/application_default_credentials.json}.
}
\examples{
\dontrun{
library(etlTurtleNesting)
library(wastdr)
library(drake)
drake::clean("wastd_reports")
drake::make(etlTurtleNesting::wastd_reports(), lock_envir = FALSE)

drake::loadd("wastd_data")
}
}
\seealso{
https://gargle.r-lib.org/articles/how-gargle-gets-tokens.html#credentials_app_default
}
