% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notify_teams_data.R
\name{notify_teams_data}
\alias{notify_teams_data}
\title{Notify a Team's channel about refreshed data.}
\usage{
notify_teams_data(team_name, channel_name = "Data")
}
\arguments{
\item{team_name}{An existing Team name, e.g. "TurtleData", "TurtlesRioTinto",
"TurtlesC4H", "TurtlesPendoley".}

\item{channel_name}{An existing Teams channel name, default: "Data".}
}
\description{
Notify a Team's channel about refreshed data.
}
