% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_tagexplorer.R
\name{update_tagexplorer}
\alias{update_tagexplorer}
\title{Update the Google Sheet "Turtle Tag Explorer"}
\usage{
update_tagexplorer(wastd_data, doc_id = Sys.getenv("TURTLE_SHEET"))
}
\arguments{
\item{wastd_data}{The output of \code{wastdr::download_wastd_turtledata()}..}

\item{doc_id}{The Google Sheet ID, default: \code{Sys.getenv("TURTLE_SHEET")}.}
}
\value{
A list with keys:
\itemize{
\item \code{tags} The selected columns and (all) rows of \code{wastd_data$turtle_tags}.
\item \code{tag_sheet} The Google sheet ID for the sheet containing \code{tags}.
\item \code{meta_sheet} The Google sheet ID for the sheet containing the metadata.
}
}
\description{
This function first authenticates (out of band) against a Google account
given through the email \code{Sys.getenv("GOOGLE_EMAIL")}.
Next, selected columns of \code{wastd_data$turtle_tags} are written over an
existing sheet "Tags" in the Google Sheet with ID
\code{Sys.getenv("TURTLE_SHEET")}.
Lastly, metadata (number of records, data accessed on, data uploaded on) are
written to a sheet "Metadata".
Messages choose verbosity from \code{wastdr::get_wastdr_verbose()}.
}
\details{
Should the total number of cells exceed the Google Sheet limit of 5M cells,
this function aborts with an informative error message.
}
