% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_sharepoint.R
\name{upload_to_sharepoint}
\alias{upload_to_sharepoint}
\title{Upload all outputs to various SharePoint sites}
\usage{
upload_to_sharepoint(wastd_reports = NULL)
}
\arguments{
\item{wastd_reports}{The result from \code{generate_wastd_reports()}
is used here to declare a dependency of this step to the reports being
generated.}
}
\description{
Refresh tokens:
}
\details{
See https://github.com/Azure/Microsoft365R/blob/master/vignettes/auth.Rmd
If token expires, run the next two lines by hand:

options(microsoft365r_use_cli_app_id=TRUE)
teams <- list_teams()

Reset tokens:

AzureAuth::clean_token_directory()
AzureGraph::delete_graph_login(tenant="dbca")
}
