% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_to_sharepoint.R
\name{upload_to_sharepoint}
\alias{upload_to_sharepoint}
\title{Upload all outputs to various SharePoint sites}
\usage{
upload_to_sharepoint()
}
\description{
Upload all outputs to various SharePoint sites
}
