% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wastd_reports.R
\name{wastd_reports}
\alias{wastd_reports}
\title{Generate and disseminate WAStD reports}
\usage{
wastd_reports()
}
\description{
Generate and disseminate WAStD reports
}
\examples{
\dontrun{
library(etlTurtleNesting)
library(wastdr)
library(drake)
drake::clean("wastd_reports")
drake::make(etlTurtleNesting::wastd_reports(), lock_envir = FALSE)

drake::loadd("wastd_data")
}
}
